user profile changes

+++ src/main/java/com/ontarget/api/dao/UserDAO.java
@@ -6,7 +6,7 @@ import com.ontarget.bean.UserDTO;
  * Created by sumit on 12/2/14.
  */
 public interface UserDAO extends GenericDAO<UserDTO> {
-    UserDTO getUser(long userId) throws Exception;
+    UserDTO getUser(Integer userId) throws Exception;
 
     public int saveForgotPasswordRequest(int userId, String forgotPasswordToken) throws Exception;
